backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM (OpenAI-compatible client)
openai>=1.50.0

# DashScope (Alibaba Cloud - API providers for ASR & TTS)
dashscope>=1.25.6

# Audio processing
numpy>=1.24.0

# Database (async PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
alembic>=1.14.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
email-validator>=2.0.0
google-auth>=2.23.0
slowapi>=0.1.9
httpx>=0.27.0
stripe>=11.6.0
googletrans>=4.0.2
azure-cognitiveservices-speech>=1.41.0
